refactoring get_feature_names_out
