+++ skrub/_apply_to_cols.py
@@ -123,30 +123,6 @@ class SingleColumnTransformer(BaseEstimator):
                 wrapped = _wrap_add_check_single_column(getattr(subclass, method))
                 setattr(subclass, method, wrapped)
 
-    def get_feature_names_out(self, input_features=None):
-        """Return a list of features generated by the transformer.
-
-        Each feature has format ``{input_name}_{n_component}`` where ``input_name``
-        is the name of the input column, or a default name for the encoder, and
-        ``n_component`` is the idx of the specific feature.
-
-        Parameters
-        ----------
-        input_features : None
-            The input features. Ignored, only here for compatibility.
-
-        Returns
-        -------
-        list of str
-            The list of feature names.
-        """
-        check_is_fitted(self, "n_components_")
-        num_digits = len(str(self.n_components_ - 1))
-        return [
-            f"{self.input_name_}_{str(i).zfill(num_digits)}"
-            for i in range(self.n_components_)
-        ]
-
 
 def _wrap_add_check_single_column(f):
     # as we have only a few predefined functions to handle, using their exact
