+++ skrub/_text_encoder.py
@@ -413,3 +413,27 @@ class TextEncoder(SingleColumnTransformer, TransformerMixin):
                 del state[prop]
 
         return state
+
+    def get_feature_names_out(self, input_features=None):
+        """Return a list of features generated by the transformer.
+
+        Each feature has format ``{input_name}_{n_component}`` where ``input_name``
+        is the name of the input column, or a default name for the encoder, and
+        ``n_component`` is the idx of the specific feature.
+
+        Parameters
+        ----------
+        input_features : None
+            The input features. Ignored, only here for compatibility.
+
+        Returns
+        -------
+        list of str
+            The list of feature names.
+        """
+        check_is_fitted(self, "n_components_")
+        num_digits = len(str(self.n_components_ - 1))
+        return [
+            f"{self.input_name_}_{str(i).zfill(num_digits)}"
+            for i in range(self.n_components_)
+        ]
